Moved the pivoted Cholesky factorization to a new file pyscf/lib/scipy_helper.py. Replacement in scf.addons.
