+++ pyscf/lib/__init__.py
@@ -23,11 +23,13 @@ from pyscf.lib import parameters
 param = parameters
 from pyscf.lib import numpy_helper
 from pyscf.lib import linalg_helper
+from pyscf.lib import scipy_helper
 from pyscf.lib import logger
 from pyscf.lib import misc
 from pyscf.lib.misc import *
 from pyscf.lib.numpy_helper import *
 from pyscf.lib.linalg_helper import *
+from pyscf.lib.scipy_helper import *
 from pyscf.lib import chkfile
 from pyscf.lib import diis
 from pyscf.lib.misc import StreamObject
